move forta config to sdk
